requirements.txt
pandas==2.2.2
pyarrow==16.1.0
requests==2.32.3
python-dotenv==1.0.1
azure-storage-file-datalake==12.22.0
azure-identity==1.18.0
duckdb==1.1.0
google-cloud-bigquery==3.25.0
jupyter==1.0.0
matplotlib==3.9.2
seaborn==0.13.2
